Add utility method for hashing a string
The salt reset everytime VS/IIS kills the app process. This was done
as a means to allow the app to mitigate against reusing old keys.

This isn't the best way to mitigate malicious requests, but I better
have something than nothing.

Also, this is a school project. If they need to use ours as the main
app, then they've got to pay for a better service :)