Fix bug with User Account being unsuspended during account edit
I slight oversight on my part when updating a user account. Being that
the default value of a bool is false; during an account update, the
flag for suspended account will be set to false. Even though it wasn't
meant to be touched.

Hence, this new method is needed to update the appropriate fields.